fix fill mat && INT_MAX line_N

+++ inc/pathfinder.h
@@ -19,14 +19,14 @@ typedef struct s_file {
 typedef struct s_matrix {
     int count_islands;
     int **matrix;
-    double **final_matrix;
+    long **final_matrix;
     char **unique_isl;
 }              t_matrix;
 
 t_file *init_start(char *argv[]);
 t_matrix *init_matrix(t_file *data);
 t_matrix *fill_matrix (t_file *data, t_matrix *mat);
-double **floyd_algo(t_matrix *mat);
+long ** floyd_algo(t_matrix *mat);
 bool mx_all_errors(t_file *data);
 bool invalid_argc(int argc);
 bool not_exist(char *argv[]);
